% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{default_parameter_table}
\alias{default_parameter_table}
\title{Table of default parameters}
\description{
A table of the default parameters entered to \code{\link{compute_design_performance}}.
}
\details{
Each row of this table corresponds to one input to \code{\link{compute_design_performance}}, and contains: (1) the name of the argument passed to \code{\link{compute_design_performance}}, (2) a label describing the input in more detail, (3) the minimum value the input can take, (4) the maximum value the input can take, and (5) the input's default value.
}

